Alterações no Domain

+++ IMDb.Server.Domain/Entities/Movies.cs
@@ -5,7 +5,7 @@ namespace IMDb.Server.Domain.Entities;
 public class Movies : Entity
 {
     public int TotalVotes { get; set; } = 0;
-    public decimal Rating { get; set; } = 0;
+    public double Rating { get; set; } = 0;
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public DateTime ReleaseDate { get; set; }
